Add AStar algorithm
The heuristic function calcultes the row and column difference
of 1~4 between current state and goal state as the cost-to-go.

This is not a perfect heuristic (not sure), therefore cannot guarantee optimal
solution.

The reason why 1~8 is not all used is because the cost often
exceed the true cost_to_goal.